Update README.md
Provided a full writeup of how to use this script.  Did not include possible pitfalls of removing an https binding from a site that uses this certificate, as that is a bug with IIS (short version: it will break all other sites; just fix the binding on any other site and it will fix all of them; applicable to Server 2008 R2).